pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-flink>=2.1.0",
    "apache-flink-statefun>=3.3.0",
    "pip>=25.2",
    "pyrex-rocksdb>=0.1.4",
    "rocksdb-py>=0.0.6",
    "watchdog>=6.0.0",
]
